Made Changes

+++ unidentified_bodies/models.py
@@ -20,7 +20,7 @@ class UnidentifiedBody(models.Model):
         ('dark_skinned', 'Dark Skinned')
     ]
 
-    id = models.SmallIntegerField(primary_key=True)  
+    id = models.AutoField(primary_key=True)  
     staff_id = models.ForeignKey(MortuaryStaff, on_delete=models.CASCADE)  
     name = models.CharField(max_length=50)  
     gender = models.CharField(max_length=50)
